+++ netket/sampler/__init__.py
@@ -1,13 +1,17 @@
 from .abstract_sampler import AbstractSampler
-from .metropolis_hastings import MetropolisHastings
-from .metropolis_hastings_pt import MetropolisHastingsPt
-from .metropolis_local import MetropolisLocal, MetropolisLocalPt
-from .metropolis_exchange import MetropolisExchange, MetropolisExchangePt
-from .metropolis_hamiltonian import MetropolisHamiltonian, MetropolisHamiltonianPt
-from .custom_sampler import CustomSampler, CustomSamplerPt
-from .exact_sampler import ExactSampler
 
+from . import numpy
 from ..utils import jax_available
 
 if jax_available:
     from . import jax
+
+from .metropolis_hastings import MetropolisHastings, MetropolisHastingsPt
+
+# from .metropolis_hastings_pt import MetropolisHastingsPt
+# from .metropolis_local import MetropolisLocal, MetropolisLocalPt
+from .metropolis_exchange import MetropolisExchange, MetropolisExchangePt
+
+# from .metropolis_hamiltonian import MetropolisHamiltonian, MetropolisHamiltonianPt
+# from .custom_sampler import CustomSampler, CustomSamplerPt
+# from .exact_sampler import ExactSampler
