+++ netket/sampler/jax/__init__.py
@@ -1,3 +1,6 @@
 from .metropolis_hastings import MetropolisHastings
 from .metropolis_local import MetropolisLocal
-from .metropolis_exchange import MetropolisExchange
+from .metropolis_exchange import *
+
+
+from ...machine import Jax as JaxMachine
