+++ netket/sampler/jax/metropolis_exchange.py
@@ -1,9 +1,8 @@
 import jax
-from .metropolis_hastings import MetropolisHastings
 import numpy
 
 
-class _exchange_kernel:
+class ExchangeKernel:
     def __init__(self, hilbert, d_max):
         clusters = []
         distances = jax.numpy.asarray(hilbert.graph.distances())
@@ -37,63 +36,3 @@ class _exchange_kernel:
 
     def random_state(self, key, state):
         return key, jax.numpy.asarray(self._hilbert.random_vals())
-
-
-class MetropolisExchange(MetropolisHastings):
-    r"""
-        This sampler acts locally only on two local degree of freedom :math:`s_i` and :math:`s_j`,
-        and proposes a new state: :math:`s_1 \dots s^\prime_i \dots s^\prime_j \dots s_N`,
-        where in general :math:`s^\prime_i \neq s_i` and :math:`s^\prime_j \neq s_j`.
-        The sites :math:`i` and :math:`j` are also chosen to be within a maximum graph
-        distance of :math:`d_{\mathrm{max}}`.
-
-        The transition probability associated to this sampler can
-        be decomposed into two steps:
-
-        1. A pair of indices :math:`i,j = 1\dots N`, and such
-           that :math:`\mathrm{dist}(i,j) \leq d_{\mathrm{max}}`,
-           is chosen with uniform probability.
-        2. The sites are exchanged, i.e. :math:`s^\prime_i = s_j` and :math:`s^\prime_j = s_i`.
-
-        Notice that this sampling method generates random permutations of the quantum
-        numbers, thus global quantities such as the sum of the local quantum numbers
-        are conserved during the sampling.
-        This scheme should be used then only when sampling in a
-        region where :math:`\sum_i s_i = \mathrm{constant}` is needed,
-        otherwise the sampling would be strongly not ergodic.
-    """
-
-    def __init__(self, machine, d_max=1, n_chains=8, sweep_size=None):
-        r"""
-        Args:
-              machine: A machine :math:`\Psi(s)` used for the sampling.
-                       The probability distribution being sampled
-                       from is :math:`F(\Psi(s))`, where the function
-                       :math:`F(X)`, is arbitrary, by default :math:`F(X)=|X|^2`.
-
-              d_max: The maximum graph distance allowed for exchanges.
-              n_chains: The number of Markov Chain to be run in parallel on a single process.
-              sweep_size: The number of exchanges that compose a single sweep.
-                          If None, sweep_size is equal to the number of degrees of freedom (n_visible).
-
-
-        Examples:
-              Sampling from a RBM machine in a 1D lattice of spin 1/2, using
-              nearest-neighbours exchanges.
-
-              >>> import netket as nk
-              >>>
-              >>> g=nk.graph.Hypercube(length=10,n_dim=2,pbc=True)
-              >>> hi=nk.hilbert.Spin(s=0.5,graph=g)
-              >>>
-              >>> # RBM Spin Machine
-              >>> ma = nk.machine.RbmSpin(alpha=1, hilbert=hi)
-              >>>
-              >>> # Construct a MetropolisExchange Sampler
-              >>> sa = nk.sampler.MetropolisExchange(machine=ma)
-              >>> print(sa.machine.hilbert.size)
-              100
-        """
-        super().__init__(
-            machine, _exchange_kernel(machine.hilbert, d_max), n_chains, sweep_size,
-        )
