+++ netket/sampler/numpy/metropolis_hastings_pt.py
@@ -1,7 +1,7 @@
 import numpy as _np
-from .abstract_sampler import AbstractSampler
+from ..abstract_sampler import AbstractSampler
 
-from ..stats import mean as _mean
+from ...stats import mean as _mean
 
 from numba import jit
 from numba import int64, float64
@@ -10,45 +10,9 @@ import math
 
 
 class MetropolisHastingsPt(AbstractSampler):
-    """
-    ``MetropolisHastingsPt`` is a generic Metropolis-Hastings sampler using
-    a local transition kernel to perform moves in the Markov Chain and replica-exchange moves
-    to increase mixing times.
-    The transition kernel is used to generate a proposed state :math:`s^\prime`,
-    starting from the current state :math:`s`, as in ``MetropolisHastings``.
-    Each replica is at an inverse temperature :math:`\beta` linearly spaced in the interval [0,1],
-    where :math:`beta=1` corresponds to the target probability distribution,
-    and :math:`beta=0` is the flat distribution.
-    """
-
     def __init__(
-        self, machine, kernel, n_replicas=32, sweep_size=None, batch_size=None,
+        self, machine, kernel, n_replicas=32, sweep_size=None,
     ):
-        """
-        Constructs a new ``MetropolisHastingsPt`` sampler given a machine and
-        a transition kernel and a number of replicas.
-
-        Args:
-            machine: A machine :math:`M(s)` used for the sampling.
-                    The probability distribution being sampled
-                    from is :math:`P(s)=|M(s)|^p`, where the power :math:`p`
-                    is arbitrary, and by default :math:`p=2`.
-            kernel: A kernel to generate random transitions from a given state as
-                    well as uniform random states.
-                    This must have two methods, `random_state` and `transition`.
-                    `random_state` takes an input state (in batches) and
-                    changes it in-place to a valid random state.
-                    `transition` takes an input state (in batches) and
-                    returns a batch of random states obtained transitioning from the initial state.
-                    `transition` must also return an array containing the
-                    `log_prob_corrections` :math:`L(s,s^\prime)`.
-            n_replicas (int): The number of replicas used for replica-exchange moves. Each replica samples
-            sweep_size (int): The number of exchanges that compose a single sweep.
-                        If None, sweep_size is equal to the number of degrees of freedom (the input size of the machine).
-            batch_size (int): The batch size to be used when calling log_val on the given Machine.
-                        If None, batch_size is equal to the number of replicas (n_replicas).
-
-        """
         super().__init__(machine)
 
         self.n_replicas = n_replicas
