+++ setup.py
@@ -169,6 +169,7 @@ setup(
         "netket.machine",
         "netket.machine.density_matrix",
         "netket.sampler",
+        "netket.sampler.numpy",
         "netket.sampler.jax",
         "netket.stats",
         "netket.operator",
